third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/1.1/form-action-src-redirect-blocked-in-new-window-expected.txt

CONSOLE ERROR: Refused to send form data to 'http://127.0.0.1:8000/resources/redirection-response.php?host=localhost:8000&status=302&target=/security/resources/post-done-to-opener.html' because it violates the following Content Security Policy directive: "form-action 127.0.0.1:8000".

This is a testharness.js-based test.
Harness: the test ran to completion.
